Moved photo handler from app.py to main.py.
Photos uploaded will continued to be stored in existing bucket in Google Cloud Storage. A Photo Data dictionary containing unique file name (key) and photo URL (value) will be stored as a list within the Experiences collection. When a GET request is sent for images, it returns a response containing a list of all URL's under an experience.